Move experimental constants to its own location outside of the core module
