+++ experimental/table/base/pom.xml
@@ -17,20 +17,17 @@
     <artifactId>alluxio-experimental</artifactId>
     <version>2.1.0-SNAPSHOT</version>
   </parent>
-  <artifactId>alluxio-experimental-table</artifactId>
-  <packaging>pom</packaging>
-  <name>Alluxio Experimental - Table</name>
-  <description>Table utility in Alluxio Experimental Module</description>
-
-  <modules>
-    <module>base</module>
-    <module>server</module>
-    <module>client</module>
-  </modules>
+  <artifactId>alluxio-experimental-base</artifactId>
+  <packaging>jar</packaging>
+  <name>Alluxio Experimental - Base Module</name>
+  <description>Base module in Alluxio experimental modules</description>
 
   <properties>
-    <!-- These need to be defined here as well as in the parent pom so that mvn can run
-         properly from sub-project directories -->
-    <build.path>${project.parent.parent.basedir}/build</build.path>
+    <!-- The following paths need to be defined here as well as in the parent pom so that mvn can -->
+    <!-- run properly from sub-project directories -->
+    <build.path>${project.parent.basedir}/build</build.path>
+    <failIfNoTests>false</failIfNoTests>
   </properties>
 </project>
+
+
