+++ experimental/table/client/pom.xml
@@ -84,6 +84,11 @@
       <artifactId>alluxio-core-transport</artifactId>
       <version>${project.version}</version>
     </dependency>
+    <dependency>
+      <groupId>org.alluxio</groupId>
+      <artifactId>alluxio-experimental-base</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
     <!-- Internal test dependencies -->
   </dependencies>
