+++ experimental/table/server/master/src/main/java/alluxio/master/catalog/CatalogMasterFactory.java
@@ -11,7 +11,7 @@
 
 package alluxio.master.catalog;
 
-import alluxio.Constants;
+import alluxio.experimental.Constants;
 import alluxio.master.CoreMasterContext;
 import alluxio.master.MasterFactory;
 import alluxio.master.MasterRegistry;
